pythone-mqtt/requirements.txt
paho-mqtt
Flask